fixes to string ids in admin panel

+++ sites/www/src/routes/(admin)/admin/content/+page.server.ts
@@ -3,7 +3,7 @@ import { contentService } from '$lib/server/db/services/content';
 import { fail } from '@sveltejs/kit';
 
 export const load: PageServerLoad = async () => {
-	const result = await contentService.get_content_items();
+	const result = await contentService.get_content_items({ limit: 10, page: 0 });
 
 	if (!result.success) {
 		fail(400, { message: 'Error getting content' });
@@ -15,7 +15,12 @@ export const load: PageServerLoad = async () => {
 export const actions = {
 	delete: async ({ request }) => {
 		const formData = await request.formData();
-		const id = Number(formData.get('id'));
+		const id = formData.get('id') as string;
+
+		if (!id) {
+			return
+		}
+
 		try {
 			await contentService.delete_content(id);
 			return { success: true };
